appRankingの修正とrankingDesignのコンポーネントの追加
